libs/indoxArcg/requirements.txt
latex2markdown==0.2.1
loguru==0.7.2
numpy==2.0.0
pandas==2.0.3
protobuf==5.27.2
pydantic==2.8.2
PyPDF2==3.0.1
python-dotenv==1.0.1
Requests==2.32.3
setuptools==69.1.1
tenacity==8.2.3
tiktoken==0.6.0
tokenizers==0.15.2
umap_learn==0.5.6
unstructured==0.15.8
nltk==3.9.1
pillow_heif==0.18.0
